# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API tra cứu kết quả xổ số kiến thiết Việt Nam.",
        "title": "KetQuaXoSo API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/check": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lấy toàn bộ kết quả xổ số của một tỉnh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lottery"
                ],
                "summary": "Lấy kết quả xổ số",
                "parameters": [
                    {
                        "description": "Thông tin tỉnh",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CheckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.CheckResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.CheckResponse"
                        }
                    }
                }
            }
        },
        "/api/check-ticket": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Kiểm tra xem vé số có trúng thưởng hay không",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lottery"
                ],
                "summary": "Kiểm tra vé số",
                "parameters": [
                    {
                        "description": "Thông tin vé số",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.TicketResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.TicketResponse"
                        }
                    }
                }
            }
        },
        "/api/province": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Trả về danh sách các tỉnh có hỗ trợ xổ số",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Province"
                ],
                "summary": "Lấy danh sách tỉnh",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CheckRequest": {
            "type": "object",
            "properties": {
                "province": {
                    "type": "string",
                    "example": "Lâm Đồng"
                }
            }
        },
        "api.CheckResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "results": {}
            }
        },
        "api.TicketRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "07/09"
                },
                "number": {
                    "type": "string",
                    "example": "123456"
                },
                "province": {
                    "type": "string",
                    "example": "Lâm Đồng"
                }
            }
        },
        "api.TicketResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "match": {
                    "type": "string"
                },
                "prize": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Nhập API key để sử dụng API. Giá trị mặc định cho test: **YouAPIKey**",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}